+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.10.1 - 28/01/2026
+* Correção em calculo de descontos em porcentagem.
+
+# 2.10.0 - 13/01/2026
+* Adição de configuração para definir tempo de expiração do código OTP;
+* Adição de configuração para definir valor mínimo de doação.
+
 # 2.9.1 - 06/01/2026
 * Correção de erro no valor dos produtos da fatura.
 * Correção em carregamento de scripts de orçamento.